requirements.txt
streamlit 
google-generativeai 
langchain 
faiss-cpu 
python-dotenv 
langchain-google-genai
langchain-text-splitters
langchain-community
langchain-core
sentence-transformers